[chore] 파싱 리소스 인터페이스 분리
